ok better handling update..
yasb doesn't need to be killed now for fullscreen windows.. - optionally still can be killed and launched with alt+ctrl+b

gaps adjusted

alt+ctrl+f fullscreens an application and launches powershell in the background to execute a redraw to fix gaps issue when going out of fullscreen with some applications... This is a dumb fix tbh and should just be solved by the GlazeWM devs but gaps remain broken after a toggle-fullscreen and I'm done waiting for a patch.

Initially I was thinking about going back to zebar.. but that configuration and setup is absolutely obsurd.. no thank you!